+++ frontend/app/.server/routes/helpers/protected-application-intake-family-route-helpers.ts
@@ -1,5 +1,7 @@
 import { redirect } from 'react-router';
 
+import validator from 'validator';
+
 import { createLogger } from '~/.server/logging';
 import { getAllowedTypeOfApplication, maritalStatusHasPartner } from '~/.server/routes/helpers/base-application-route-helpers';
 import { getChildrenState, getContextualAgeCategoryFromDate, getProtectedApplicationState } from '~/.server/routes/helpers/protected-application-route-helpers';
@@ -151,11 +153,7 @@ export function validateProtectedApplicationFamilyStateForReview({ params, state
     throw redirect(getPathById('protected/application/$id/intake-family/contact-information', params));
   }
 
-  if (email === undefined) {
-    throw redirect(getPathById('protected/application/$id/intake-family/contact-information', params));
-  }
-
-  if (email && !emailVerified) {
+  if (!email || !validator.isEmail(email) || !emailVerified) {
     throw redirect(getPathById('protected/application/$id/intake-family/contact-information', params));
   }
